+++ CHANGELOG.md
@@ -7,7 +7,7 @@ unreleased
 - Update browser-detection to v1.5.0
 - Add `aria-label` attribute to payment options
 - Update checkout.js to v4.0.95
-- Add `clearActivePaymentMethod` to remove selected payment method
+- Add `clearSelectedPaymentMethod` to remove selected payment method
 
 1.4.0
 ------